services/gateway/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
websockets==15.0.1
pydantic-settings==2.10.1
redis==6.4.0
psycopg[binary]==3.2.9
psycopg-pool==3.2.6
boto3==1.40.18
prometheus-client==0.22.1
PyJWT==2.10.1
python-multipart==0.0.20
